+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+# Release 2023-08-30 (2.7.5)
+
+  * Added folder icons for chart and shape libraries
+  * Added feature to make PPT Thumbnails clickable in presentation mode (hyperlink to slide)
+  * Added feature to replace all PPT Thumbnails with regular picture (remove all meta data)
+  * Added feature to toggle and remove PPT Notes from all selected slides
+  * Added new options to style harvey moons
+  * Font icons now support Segoe Fluent Icon Font
+  * Font Awesome updated to 6.4.2
+  * Fixed icons mixed up in arrage easy group (thanks to robertzb)
+  * Fixed adding multiple shapes to a shape library didn't work
+  * Fixed umlauts in install path / config paths didn't work
+  * Fixed handling of animations in group helper functions and state-shapes
+
+
 # Release 2022-09-03 (2.7.4)
 
   * Restructuring of context menus to load various function dynamically based on selected shape